UTs: Fix clearing of RPC exchange between tests
Our current code does not properly clean up the fake RPC exchange between
tests.

This means that if a test invokes code in cinder that makes an RPC
request and it is not consumed by that test, then another test may get
some code invoked.

One test that makes such an RPC request that is not consumed is our test
cinder.tests.unit.policies.test_snapshots.SnapshotsPolicyTest.test_delete_policy_6_project_reader

This means that if the next tests starts the volume manager service and
asserts that cinder.volume.volume_utils.notify_about_snapshot_usage is
not being called, it will fail, as the notification would have been
called when processing the previous test deletion request.

This was found when working on Change-Id patch
I60527edb029265ed4b2a48d2c358265cea1a9b2c and analyzing a spurious
failure and finding out that running these 2 tests one after the other
was the problem:

  cinder.tests.unit.policies.test_snapshots.SnapshotsPolicyTest.test_delete_policy_6_project_reader
  cinder.tests.unit.test_volume_transfer.VolumeTransferTestCase.test_transfer_accept_with_snapshots_1_True

The second test is added in that Change-Id that follows this one in the
patch dependency chain.

This patch fixes this issue by forcefully clearing the class dictionary
in the fake exchange.

We do it forcefully because the oslo messaging library doesn't have any
other mechanism to do so for now, but once patch
https://review.opendev.org/c/openstack/oslo.messaging/+/901018 is merged
we can remove this code.

Change-Id: I52ee4b345b0a4b262e330a9a89552cd216eafdbe