[SAP] handle DuplicateName in _migrate_attached_cross_vc
The diskless backing could have already been created by a
previous (failed) migration attempt. We do not clean up the
diskless backings for failed migrations, and thus we might
end up with DuplicateName error in subsequent attempts.

We attempt to reuse the existing backing by checking if it
is diskless (doesn't have any disks attached to it). Otherwise
we let cinder to update the host of the volume, but set the
migration_status to error.

Change-Id: I2e0bd27b649bd75f988a5661f83ff476020dc49b