Change log message in get_qemu_data
The get_qemu_data method in image_utils file raises an exception when
disk_format is not of type RAW. The message supposed to say what disk type
are using, but return False instead.

So i change the log message to something more general, just specifying that
the disk are being used is not of RAW type.

Closes-Bug: #2008041
Change-Id: I5c73b85531897656cb6a9d55391933ce40eb0635