Don't retry service update in report_state
The service_update DB API uses the oslo.db
retry logic with increasing retry interval
for each retry.

Due to the above if you for example block
the database connection the retry logic will
block the report for an extended period of
time that causes the service to be down even
if the service has recovered.

The retry logic is not optimal for the
functionality that report_state does where
we want it to fail fast and try report again.

Alternatives to this patch is duplicating the
service_update() DB API logic or remove the
wrapped retry on service_update completely but
that will also affect other parts of the code.

Closes-Bug: 2018595
Change-Id: I557d0eca5d8e85b15419591119758c635fb51666