Run protection tests during gate check
This patch adds the cinder-tempest-plugin-protection-functional test to
the gate check.  The test itself is defined in cinder-tempest-plugin.

This test enables Consistent and Secure Default RBAC (SRBAC) [1] on both
Keystone and Cinder to ensure that integration with Keystone and Cinder
functionality is working correctly when both services are enforcing
scope in RBAC.

Currently no other tests in the Cinder gate have this option enabled.

Moving forward, this job will help ensure that new changes to Cinder
will work in environments where SRBAC is enabled.  This job will also
help as we continue to implement the next phases of the SRBAC community
goal.

[1] https://governance.openstack.org/tc/goals/selected/consistent-and-secure-rbac.html

Depends-On: https://review.opendev.org/c/openstack/cinder-tempest-plugin/+/906153
Change-Id: I590eb954203ccd2fc564117addfcf5d2dac5bd74