[SAP] change storage_profile reporting
This patch changes the way the storage_profile attribute is reported
in the pool stats.  Previously the value was a dictionary that
included the name and the storage profile id.   The storage_profile
needs to only be the name so we can match against it in the
volume type extra spec.

This is related to the cinder-issue:
https://github.wdf.sap.corp/cc/cinder-issues/issues/14