Fujitsu Driver: Improve CLI function
When the CLI command execution encounters an error, the storage
returns an error code.
However, only some error codes could be parsed, resulting the
inability to view specific error information in the log.
To fix this issue, the parsing of error codes has been added when
the CLI encounters an error, allowing the error information to be
viewed in the log.

In addition, this patch also adds support for SSH keys,
allowing connections to the storage using SSH key.

Closes-Bug: #2048850
Change-Id: I204f3889e3401b142dd7c9a6e11585639585abfc