Fix online data migration
The online migration remove_temporary_admin_metadata_data_migration
was recently merged and broke gate for the kolla project.

There are 2 issues with the current code:
1. The db api doesn't return the total and updated values
2. We are issuing a limit and update query together which is
not allowed generating the following error:

sqlalchemy.exc.InvalidRequestError: Can't call Query.update()
or Query.delete() when limit() has been called

This patch fixes the issue by creating a select subquery to get
all the ID values with the limit as max_count. We then create
a new query for update and pass the select subquery as a filter.

We don't require a releasenote since the bug and fix are in the
same release.

Closes-Bug: 2052805
Change-Id: Ida994f767eecb094c177db15dfc80a0c0fe56447