tests: Merge v2, v3 volume API tests
We no longer support the v2 API. Given the v3 controller is the v2
controller but with the addition of microversions and some extra routes
*and* the v2 tests obviously don't set microversions or test any of
those additional routes, we can simply point the v2 controller tests at
v3 controller and things will just work (TM). We do just this, moving
the test to the v3 test file in the process.

Change-Id: Ib67f28f422db9255053e26821f7484808d0428fc
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>