Add support for glance new location APIs
This patch adds the call for ``add_image_location`` API
which triggers the new location API workflow in glance
that addresses OSSN-0065. It is more secure and robust
compared to the old location workflow.
This call will be made when glance is using cinder as
a backend and we want to perform an optimized
``upload volume to image`` operation.

Signed-off-by: Rajat Dhasmana <rajatdhasmana@gmail.com>
Change-Id: Ib70219a9d085257c90a75ddcfcb935b4659fd28d