tests: Merge v2, v3 snapshot API tests
We no longer support the v2 API. Given the v3 controller is the v2
controller but with the addition of microversions and some extra routes
*and* the v2 tests obviously don't set microversions or test any of
those additional routes, we can simply point the v2 controller tests at
v3 controller and things will just work (TM). We do just this, moving
the test to the v3 test file in the process.

Change-Id: If90d2029fec5a5f5fce95a2ff7cbad2bdc71e495
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>