Add testing for optimized volume upload
When using cinder as glance backend, we have an optimization for
upload volume to image operation.
It was recently found that the glance location API was broken for
the upload volume optimization. The current glance cinder job
doesn't run the upload volume test hence the issue was not detected.

This patch adds the volume action tests (that includes the upload
volume test) in the glance cinder job. It also requires configuring
some parameters that are set up in devstack in the depends-on patch.

Depends-On: https://review.opendev.org/c/openstack/cinder/+/909513

Signed-off-by: Rajat Dhasmana <rajatdhasmana@gmail.com>
Change-Id: I4918027ed641a90aafe44d815a4a3dbc1dc7ddfc
(cherry picked from commit 787ca3db096fca034fb070fc77a2c55630bcea80)