Fix: user_id in clone volume
When cloning a volume from source, we overwrite some values based
on the destination volume details.
However, we use the user_id of the source volume and don't overwrite
it with the context of the current user (or service).
This patch fixes the issue by using the context's user_id instead of
the source volume's user ID.

Closes-Bug: #2056081
Change-Id: I1c3d44e108a6416d7a82fae9246a9b25a069f3eb