Ensure backup availability zone is populated if empty
If the availability zone is not provided when creating the backup,
the availability zone is still None after scheduling.

So just before starting the backup, we can fill the availability_zone
property with the actual az of the backup service doing the backup.

This will also fix selecting the proper host when restoring a backup.
If the az field is empty, it would randomly pick a cinder-backup service,
while it would make sense that only the services in the zone of the backup
would have access to the backup resource.

Closes-Bug: #1942154
Change-Id: Ida930463228bb60f71bb4b80d3f193dde8398a70
(cherry picked from commit 9b7eaf2e328400b2ad00a93d8cd602ebcce63731)