api: Keep track of extension controllers
We need to be able to resolve the original, unversioned controllers, so
that we can retrieve additional attributes of those controller (namely,
the 'versioned_methods' attributes, which contains references to the
original versioned methods rather than the 'version_select' magic
method). Register these things slightly differently. It would likely be
better to fold these action controllers into the main controllers, but
that's a lot of code motion that I don't really want to do right now.

Change-Id: I174c3bae39625b74e4b416e3559eb79ad8d34948
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>