api: Add remaining missing request body schemas
This is relatively simple. The schemas added are either (a) extremely
loose or (b) merely mimic validation that already exists in the API
methods. Where the latter occurs, we remove the existing validation
logic, which effectively means removing invocations of
'assert_valid_body' and 'validate_name_and_description'. So far, so
good. We also rename the imports for consistency as we plan to conduct
far more surgery on these.

We do, however, uncover a bug. The 'VolumeTypesManageController._delete'
method is getting mapped to two APIs: 'DELETE /types/{id}' and 'POST
/types/{id}/action'. Only the former is expected and attempting to use
the latter fails, but it fails with HTTP 500 (bad!) rather than HTTP 404
(good!). This can be reproduced with e.g.

  >>> import openstack
  >>> conn = openstack.connect()
  >>> type_id = 'xxx'  # a valid type ID here
  >>> body = {'delete': None}
  >>> conn.block_storage.post(f'/types/{type_id}/action', json=body)

We should fix this, but doing so is out-of-scope for this change so a
FIXME is included for now instead.

On top of all the above, the schema decorator we have expects a 'body'
kwarg to be passed to the method and requires a request object with API
version attributes. We had a couple of tests that were passing body as a
positional argument instead or using fake request objects. We correct
all these issues.

Change-Id: I0d4ed51932ba94abb59e650fa726c877b94656b1
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>