Fix reimage with snapshot backed image
When we create a snapshot of a server that is volume-backed,
glance creates a zero sized metadata entry of the image which
is backed by the volume snapshot.
When we try to reimage a volume with that given image, we need
to fetch the volume snapshot details from the image metadata.
Now to reimage, we need to create a new volume from snapshot and
copy the data from our snapshot-volume to our original volume
for which we use the generic revert to snapshot mechanism as
it performs the same steps.

Closes-Bug: #2062539

Change-Id: Ic4bf44c320ad53b514178ecd4d5f57f037169bfe
(cherry picked from commit ae22195df7d674ba017f5e301868522ccb21c5b4)