Remove psutil requirement
Follow up on a TODO in image_utils to replace use of psutil with
the standard library shutil, which makes it no longer a requirement
for the main cinder code.  The only driver using psutil is Quobyte,
so make it a driver extra requirement, and make some minor adjustments
to the quobyte driver and tests so that the driver will check for the
availability of psutil and the unit tests can operate without it being
present.

Change-Id: Icd1a4e59317fe1db0c5419b9aaeb85464b7be863