api: Simplify enable/disable APIs (clusters)
I'm using simplify a little liberally here: what we are actually
simplifying is our later addition of API schemas. This change migrates
us from using route's "update" method to using a proper "action" method
like we use for most other actions. This allows us to work towards the 1
schema 1 method model that is easiest to introspect.

Note that to do this, we must migrate this resource away from
'mapper.resource'. What we wanted to do was this:

  mapper.resource(
      'cluster',
      'clusters',
      controller=self.resources['clusters'],
      collection={
          'detail': 'GET',
          'enable': 'PUT',
          'disable': 'PUT',
      },
  )

However, this doesn't work because Routes has a feature/bug where if we
pass non-GET actions via the 'collection' argument, then the first of
these will be marked as the "primary" route. When this happens, we don't
use the action to lookup the controller method nor do we suffix the path
with the action [1]. In concrete terms, instead of getting a 'PUT
/clusters/enable' route that resolves to 'ClusterController.enable', we
get a 'PUT /clusters' route that resolves to a non-existent
'ClusterController.update' method. The solution to this is to use
'mapper.connect' - or more specifically a newly introduced
'mapper.resource' wrapper based on nova's equivalent [2] - to manually
specify each path instead.

[1] https://github.com/bbangert/routes/blob/v2.5.1/routes/mapper.py#L1157-L1169
[2] https://github.com/openstack/nova/blob/29.0.1/nova/api/openstack/__init__.py#L203-L215

Change-Id: I364d92738da7fecbc9c278dbeedaa74c63ee1595
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>