Integrate 'pre-commit'
This nifty tool provides a extensible way to configure pre-commit hooks
to do everything from running linters to spellchecking [1] and has
already been adopted by a few OpenStack projects [2]. Let's copy what
we've done elsewhere and migrate most of our linters to this. This makes
my life much easier as I work on the OpenAPI series since linting is
taken care of. Users who do not wish to use this can simply not do so.

[1] https://pre-commit.com/
[2] https://review.opendev.org/#/q/file:.pre-commit-config.yaml

Change-Id: I7dd14b4087d92f6bdae575581e03f991c519d57a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>