[IBM SVf Driver] Added support for restorefromsnapshot CLI for volumegroups
[Storage Virtualize family] As part of Flashcopy 2.0 implementation,
added support for restorefromsnapshot for a volumegroup using an
existing volumegroup snapshot.

Currently there is no cinder API for group-snapshot-restore but
IBM storage supports this feature.
Hence adding the capability of this CLI in the IBM SVf Driver,
so that IBM's internal customers can build their own API and
use this functionality.

Implements: blueprint ibm-svf-volumegroup
Change-Id: I3a6906583a43ceab1f3e217666768d12fdb214db
Signed-off-by: harshailani <harsh.ailani@ibm.com>