Make default-types APIs compatible with V3.67
In microversion 3.67, we made project_id optional in the URL
for all Cinder APIs.
However, the default-types APIs (set, unset, get, list) were
implemented without the project_id in the URL to make it
ready for SRBAC and anticipating the new endpoint for cinder
without project_id.

This is causing issues while implementing the SDK support for
default-types APIs since we fetch the endpoint containing
project_id (http://127.0.0.1/volume/v3/<project_id>/default-types)
but the default-types API doesn't expect it
(http://127.0.0.1/volume/v3/default-types) resulting in 404
resource not found error.

ResourceNotFound: 404: Client Error for url:
http://127.0.0.1/volume/v3/default_types/12a1b5e507e7497db79707b0ddedf1a4,
: 404 Not Found: The resource could not be found.

This patch makes the default-types APIs consistent with the
other cinder APIs that accept URL with/without project_id.

NOTE: This patch doesn't require a microversion bump or releasenote
since the expectation with MV3.67 is that all APIs support
request irrespective of project_id in the URL and this change is
making default-types APIs consisitent with MV3.67.
Also MV3.67 is just an indication of the behavior change otherwise
supplying URLs without project ID with MV < 3.67 also works.

Change-Id: I63efc0598d501d77474588a02582f5338bb8d345