Stop testing old release of cinder
A test in datera was failing locally due to the fact the the version
checking could be wrong in the egg-info local folder if you switch from
a branch to another.

It's not necessary to test this anymore as Cinder version 15 (train) is
EOL anyway.

Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>
Change-Id: I0d17899c0b239ec969ff51238dcbd4402a567f6c
(cherry picked from commit 7ee7ac5899b451a9d9e2f7b0a2b664fcd5ba3a57)