Use only absolute path with qemu-img info
We're already doing this in image_utils since we're looking at a
tempfile created by mkstemp (which returns an absolute path to the
file).  Harden the code by making this use of an absolute path
explicit, because calling qemu-img info without an absolute path
leaves us open to the CVE-2024-4467 exploit.

Change-Id: I1649384fe4d7baf665a95e05ac36384b4b4e425b
Closes-bug: #2073275
Signed-off-by: Brian Rosmaita <rosmaita.fossdev@gmail.com>