More thorough mock on a backup message test
Modify test_backup_create_scheduling_error to mock
_choose_backend_filters(), which when left unmocked eventually causes
a stevedore extension_manager to be loaded, which on some systems
cannot find the filter classes.  Since we want to test the generation
of backup user messages and not the actual loading of filters, mock
the function so that we can test what we're interested in.

Change-Id: I98052eafc3a31ae64890084e943f8454d98c5a91