Do not call get_all_classes to get all classes
The BackendFiltersTestCase creates a class map of filters to use
in its tests, but it is populated by calling get_all_classes on
a BackendFilterHandler, which ultimately creates a stevedore extension
manager to load the filter classes.  Unfortunately, on some systems
the extension_manager cannot find the classes.  Since what we want
to test in this file is that the filters work as expected, not that
stevedore can find and load them, use a pre-populated map instead.

Change-Id: I36ce6aee924f3aea6f0ec0f87831db4ddf4b5753