Remove hard coded class in FakeFilterScheduler
The FakeFilterScheduler was using a hard-coded class to recreate the
host_manager that had already been injected into it when it was
initialized.  Given that the injection is controlled by the config
opt 'scheduler_host_manager', this hard coding is counterproductive,
so remove it.

Additionally, some filter scheduler tests were re-creating a
scheduler's host_manager yet a third time to set it to a
FakeHostManager; this is no longer necessary as the superclass
of the FilterSchedulerTestCase is already setting the
scheduler_host_manager to FakeHostManager during setUp(), and that
setting is now no longer overriden.

Change-Id: Ib32977d032c62b5b607ed5a30cc78c2288c2d7bd