Mock get_all_classes in setUp()
On some systems, many host manager tests are failing during the
setup phase when the host manager is initialized with a hard-coded
backend filter handler that ultimately uses stevedore to locate the
filter classes, but can't find them.  Since we don't want to test
stevedore in these unit tests, mock the backend filter handler
get_all_classes function so initialization of the host manager
can complete and the unit tests can be executed.

Also add this mock to a few tests that create their own host manager
instead of relying on the ones created in setUp().

Change-Id: I709b1755c769dedeaeed193106fb84cd4a4b3b57