Skip image format detection tests for compute
The ImagesFormatTest was added to tempest to check
that nova is correctly rejecting possibly malicious
images.

In the cinder-plugin-ceph-tempest job, however,
openstack is configured with an optimized storage
path and hence nova doesn't examine the
image (as it would do in the "normal" path), and
the expected exceptions are not raised, leading
to test failures. So configure the
cinder-plugin-ceph-tempest job to skip these
tests until the nova team figures out how to
handle this.

This patch is based on https://review.opendev.org/c/openstack/nova/+/926215

Co-Authored-By: Brian Rosmaita <rosmaita.fossdev@gmail.com>
Change-Id: I43ef463ca8ec17613b56d75b94b2d3dd81054280