[Pure Storage] Fix CG, FC WWN and NVME-TCP NSID bugs
It appears that the format of the WWN returned from a FlashArray
has changed so we need to specifically remove colons from the
WWN to allow os-brick to connect to the correct fabric port.

From Purity 6.6.0 NVMe-TCP volumes now return an NSID instead
of a LUN ID. Added additional code to use the correct value
depending on the Purity version.

Fix an issue where a volume not in a protection group causes
a consistency group clone to fail because of an AttributeError.

Fix create from CG snapshot issue caused by cgsnapshot value
not being populated into snapshot dict.

Change-Id: I4ff2ee20e4ab82577a3f3cb9ccd07e468981a6f4