NetApp: NVMe namespace mapping fails during VM live migration
When new Cinder volume is created a new subsystem and
NVMe namespace created and mapped one-to-one in ONTAP.
Attach process will continue this one-to-one mapping
between subsystem and host. This approach limits
NVMe multi attach workflow and Live migration.
Single subsystem mapped to single namespace is causing
this limitation. ONTAP dosn't allow mapping
multiple subsystems to single namespace. This contrasts
with iSCSI, where multiple iGroups can be mapped
to a single LUN. To overcome this limitation,
the workflow is changed to map multiple hosts to
a single subsystem. Each subsystem then has a
one-to-one mapping with a namespace.

Closes-bug: #2078968
Change-Id: I90b9d74560f128e495b6ccecea2d3aa2ed68171f
Signed-off-by: jayaanand borra <jayaanand.borra@netapp.com>