datacore: Mock is_valid_ipv6 to fix unit tests with netaddr >= 1.0.0
netaddr 1.0.0 changed its behavior and it now raises TypeError in case
a non string value is passed. Now the exception is not caught in
oslo.utils, and we may not catch it there because using the function
for non-string value is apparently incorrect.

Mock the method in datacore unit tests which does not present strings
to the function because of quite "naive" mocking. We can later drop
the mock if anyone from Datacore can work on updating the mock strategy
in their unit tests.

Change-Id: Ibe4d4df6915dd621a13c2108f5d4b3194498ce55