Fix: Unset default type for non-existent project
When performing unset operation of a default type for a particular
project, if the project doesn't exist then we fail the operation.
This makes it impossible to remove the default type record (unless
we perform manual DB removal) and it shows up in the default type
list forever.
This patch addresses the issue by ignoring the project not found
error and proceeding with the unset operation anyway.

Closes-Bug: #2081283
Change-Id: I5c4b0907f65e3fa9d9589101a2de04309cfe7fbf