Add 'enforce_multipath' in connection_properties
Previously, the check for multipath used to happen while fetching
connector information. The downside of this was we didn't know
what kind of connection (iSCSI, FC, NVMe) will be created and
it was not possible to handle the multipath checks for different
connectors.
With change[1], the check was moved to the connect_volume method
but that led to another problem that the connection information
returned from cinder didn't contain the ``enforce_multipath``
parameter.
The purpose of this patch is to include the ``enforce_multipath``
parameter in the connection information based on the value fetched
from the connector.

[1] https://review.opendev.org/c/openstack/os-brick/+/933250

Partial-Bug: #2085013

Change-Id: I4ea618a44b37cce5dd0b131ffac55bce51d08c0a