[CI] Unblock the gate
Recently py38 support was dropped from Zuul's ubuntu-focal nodeset.
Because of this the tempest version used so far had to be set to older
one [1] which does not have the 'slow' tox target anymore.
This causes tempest-slow-py3 to fail. Setting it to non-voting until
there isn't a better solution. Probably to use an older job definition
that is there on the chosen tempest version.

Due to the same reason tempest-itegrated-storage job fails with multi-
attach related test cases, as in xena branch the multiattach parameter
was removed [2] and the older tempest version still uses that. This
patch add exclude regex to skip the related tests.

Another issue is with cinder-plugin-ceph-tempest job which times out
during pip install due to (probably) wrong upper constraints version.
To unblock the gate this job is also set as non-voting.

[1] https://review.opendev.org/c/openstack/devstack/+/933741
[2] https://review.opendev.org/c/openstack/cinder/+/880375

Change-Id: Ia148887f8323aa71b5f0533052c2f026e4124a08