[SAP] sensitive connection_info to only be sent for os_brick
An attempt to attach a volume to a baremetal instance would
generate a connector with os_type='platform', thus we need to be
more specific when checking if the connector is os_brick.

This prevents any sensitive information that's passed to os_brick
from being saved in the attachments database.

Change-Id: I58f9c8e3f57bd632ae8abd74f9afc9acd070363c