Doc: Improve default quota value for projects
I01114211010817fbc40e959c3db9045958c78d06 intorduce the guide for
update of the default cinder quota value for a new project.
However, the actual behavior of the update affects to the existing
projects if those projects' quota value use the default value.

To elaborate the current behavior, this fix provides an exmample and
update the description.

Change-Id: I350a15a363393f65a716b32b240668f32ef2b5a0
Signed-off-by: Keigo Noha <knoha@redhat.com>