[Pure Storage] Remove SafeMode PGs from Replication Pods
When using synchronous replication, FlashArrays use pods to facilitate
this. If a backend array has SafeMode enabled, pods will be created
with a SafeMode protection group.

As SafeMode is not supported by Cinder we need to check if a SafeMode
protection group was created for a Cinder replication pod, and if so
remove this.

Change-Id: I2c3c82babf54b5f5f003a98a52ec3d1c5efcea5d