[Pure Storage] Manage Volume from GUI fails
Using the command line for volume management works with no issue, as the
method requires specifying ``id-type name``.

When using the GUI the selection of ``name`` as the ID Type actually
translates to ``source-name``, which seems to be the default value.

This patch changes references in the Pure Storage driver from ``name``
to ``source-name``.

This also means that using the CLI no longer requires the specific
selection of ``id-type name`` as the default value can now be used.

Closes-Bug: #2096801

Change-Id: I804ab5a88496b1da82c45ccabdb0fe175aacfbfb
(cherry picked from commit f0562c9bb0fca0ad9bac54ac125572747c6d5753)