NVMe-oF Target: Fix incorrect check for initiator in connector data
Drivers using the Cinder NVMe-oF target may fail if the
`initiator` (from iSCSI devices) isn't part of the connector data
populated by the os-brick.

The `initiator` information comes from the iSCSI connector and is
unrelated to NVMe targets, so validating such information isn't
correct.

When iSCSI isn't properly configured in the host and
`/etc/iscsi/initiatorname.iscsi` is missing, the iSCSI connector
fails to gather the `initiator` information and the property isn't
added to the connector data, exposing the issue.

This patch fixes this issue by checking only for NVMe-oF related
fields in the connector properties instead of iSCSI `initiator`
data.

Closes-Bug: #2105961
Change-Id: Ieab023f9df04779a33e535ca3b502f9a48f3ea13